Setup moved to separate package to not break everyones junit tests and jmh benchmarks
